Split monolithic test harness into tests/harness/ package
Extract 8 modules from the 4863-line monolith while preserving exact
behavior and CLI interface for meson:

  tests/harness/env.py       — platform detection, path constants, timeouts,
                               Wine subprocess patches (applied at import time)
  tests/harness/cache.py     — FPC RTL AST/codegen cache setup, test result
                               cache (key/check/store)
  tests/harness/sanitizers.py — valgrind/ASAN runner, PTY capture, Wine exec
  tests/harness/runner.py    — run_compiler(), COMPILER_RUNS, LINK_ARGS_BY_ASM,
                               coverage/target flag helpers
  tests/harness/reporters.py — TAPTestResult, TAPTestRunner, TimingTestResult
  tests/harness/parallel.py  — ParallelTestRunner, TAPParallelTestRunner and
                               helpers (_flatten_tests, _run_single_test_with_timeout)
  tests/harness/artifacts.py — failure artifact storage, read_file_content,
                               _signal_name_suffix
  tests/harness/discovery.py — FPC RTL unit detection, bootstrap flag builders,
                               Pascal comment/uses-clause parsing

The entrypoint do_not_run_me_directly_but_through_meson.py is now a thin
shim (~185 lines of imports/aliases) + TestCompiler class + discovery
functions. Meson invocation and CLI interface unchanged. Test counts:
baseline 1035 subtests passed = post-split 1035 subtests passed.
