+++ tests/do_not_run_me_directly_but_through_meson.py
@@ -1,531 +1,160 @@
 # THIS PROGRAM WILL NOT WORK IF YOU DO NOT COMPILE SOURCES FIRST WITH MESON
-import argparse
-import concurrent.futures
-import hashlib
-import json
-import locale
-import os
-import shutil
-import shlex
-import socket
-import select
-import subprocess
+#
+# This file is the entry point kept for meson compatibility.
+# Utility code has been split into tests/harness/ sub-modules.
+# See tests/harness/ for: env, cache, sanitizers, runner, reporters, parallel,
+# artifacts, discovery.
+
+# Ensure the project root is in sys.path so that `tests.harness` is importable
+# regardless of how this script is invoked (by meson, pytest, or directly).
 import sys
+import os as _os
+_project_root = _os.path.dirname(_os.path.dirname(_os.path.abspath(__file__)))
+if _project_root not in sys.path:
+    sys.path.insert(0, _project_root)
+
+# Standard library imports still needed by TestCompiler and discovery functions.
+import argparse
+import hashlib
+import os
+import re
+import shlex
+import shutil
+import socket
+import subprocess
 import tempfile
-import textwrap
-import threading
 import time
 import traceback
-import re
 import unittest
-import signal
 from pathlib import Path
-try:
-    import pty  # type: ignore
-    import termios  # type: ignore
-    import fcntl  # type: ignore
 
-    HAS_PTY = True
-except ImportError:  # Windows/POSIX-less runtimes
-    HAS_PTY = False
+# ---------------------------------------------------------------------------
+# Import from harness sub-modules (replaces the first ~1576 lines of
+# the old monolith).  Import order: env first (patches subprocess at
+# module load time).
+# ---------------------------------------------------------------------------
 
-WINDOWS_ABI_PLATFORMS = ("win", "cygwin", "msys", "mingw")
-PLATFORM_ID = sys.platform.lower()
-IS_WINDOWS_ABI = os.name == "nt" or PLATFORM_ID.startswith(WINDOWS_ABI_PLATFORMS)
-
-# Detect if we're running under Wine (Windows Python on Linux).
-# In CI this may be signaled either by WINE* environment variables or by
-# Meson's exe wrapper path when the Windows Python interpreter is launched
-# through `/usr/bin/wine`.
-_meson_exe_wrapper = os.environ.get("MESON_EXE_WRAPPER", "")
-IS_WINE = IS_WINDOWS_ABI and (
-    any(k.startswith("WINE") for k in os.environ)
-    or "wine" in _meson_exe_wrapper.lower()
-)
-EXE_EXT = ".exe" if IS_WINDOWS_ABI else ""
-
-_original_subprocess_run = subprocess.run
-
-def _patched_subprocess_run(args, **kwargs):
-    """Wrapper that resolves executable to absolute path on Wine."""
-    if IS_WINE and args and isinstance(args, (list, tuple)):
-        args = list(args)
-        if os.path.exists(args[0]):
-            args[0] = os.path.abspath(args[0])
-    return _original_subprocess_run(args, **kwargs)
-
-subprocess.run = _patched_subprocess_run
-
-# Wine's CreateProcess sometimes returns invalid thread handles, causing
-# _winapi.CloseHandle(ht) to raise OSError [WinError 6].  Patch CloseHandle
-# to silently ignore "invalid handle" errors so subprocess.Popen doesn't crash.
-if IS_WINE:
-    import _winapi  # type: ignore
-    _original_CloseHandle = _winapi.CloseHandle
-    def _safe_CloseHandle(handle):
-        try:
-            _original_CloseHandle(handle)
-        except OSError:
-            pass
-    _winapi.CloseHandle = _safe_CloseHandle
-
-    if hasattr(subprocess, "Handle"):
-        def _safe_handle_close(self):
-            if not self.closed:
-                self.closed = True
-                _safe_CloseHandle(self)
-
-        subprocess.Handle.Close = _safe_handle_close
-        subprocess.Handle.__del__ = _safe_handle_close
-
-    _original_popen_internal_poll = subprocess.Popen._internal_poll
-
-    def _safe_popen_internal_poll(self, *args, **kwargs):
-        try:
-            return _original_popen_internal_poll(self, *args, **kwargs)
-        except OSError:
-            return self.returncode
-
-    subprocess.Popen._internal_poll = _safe_popen_internal_poll
-
-    _original_popen_del = subprocess.Popen.__del__
-
-    def _safe_popen_del(self, *args, **kwargs):
-        try:
-            _original_popen_del(self, *args, **kwargs)
-        except OSError:
-            pass
-
-    subprocess.Popen.__del__ = _safe_popen_del
-
-# Path to the compiler executable
-# Get the build directory from the environment variable set by Meson.
-# Default to "build" for local testing.
-build_dir = os.environ.get("MESON_BUILD_ROOT", "build")
-KGPC_PATH = os.path.join(build_dir, "KGPC/kgpc.exe" if IS_WINDOWS_ABI else "KGPC/kgpc")
-TEST_CASES_DIR = "tests/test_cases"
-INPUT_DATA_DIR = TEST_CASES_DIR
-TEST_OUTPUT_DIR = "tests/output"
-PROJECT_ROOT = Path(__file__).resolve().parents[1]
-GOLDEN_AST_DIR = "tests/golden_ast"
-# Default execution timeout per compiled test program (seconds).
-# Can be overridden via environment variable KGPC_TEST_TIMEOUT for slower machines.
-try:
-    EXEC_TIMEOUT = int(os.environ.get("KGPC_TEST_TIMEOUT", "10"))
-except ValueError:
-    EXEC_TIMEOUT = 10
-
-# Timeout for compiler/linker subprocesses when running under Wine/MinGW.
-# Can be overridden via environment variables for slow environments.
-try:
-    COMPILER_TIMEOUT = int(os.environ.get("KGPC_COMPILER_TIMEOUT", "120"))
-except ValueError:
-    COMPILER_TIMEOUT = 120
-try:
-    LINK_TIMEOUT = int(os.environ.get("KGPC_LINK_TIMEOUT", "60"))
-except ValueError:
-    LINK_TIMEOUT = 60
-
-# Per-test timeout (entire test including compile, link, and run).
-# Useful for CI to detect hanging tests.
-try:
-    TEST_CASE_TIMEOUT = int(os.environ.get("KGPC_TEST_CASE_TIMEOUT", "300"))
-except ValueError:
-    TEST_CASE_TIMEOUT = 300
-
-# Number of parallel workers for test execution (0 = sequential).
-DEFAULT_PARALLEL_WORKERS = max(1, (os.cpu_count() or 1) * 5)
-try:
-    PARALLEL_WORKERS = int(
-        os.environ.get("KGPC_PARALLEL_WORKERS", str(DEFAULT_PARALLEL_WORKERS))
-    )
-except ValueError:
-    PARALLEL_WORKERS = DEFAULT_PARALLEL_WORKERS
-
-# Cap concurrent compiler subprocesses in parallel test mode to avoid
-# non-deterministic codegen failures under extreme process pressure.
-DEFAULT_COMPILER_PARALLEL_LIMIT = 1
-try:
-    COMPILER_PARALLEL_LIMIT = int(
-        os.environ.get(
-            "KGPC_COMPILER_PARALLEL_LIMIT", str(DEFAULT_COMPILER_PARALLEL_LIMIT)
-        )
-    )
-except ValueError:
-    COMPILER_PARALLEL_LIMIT = DEFAULT_COMPILER_PARALLEL_LIMIT
-COMPILER_PARALLEL_LIMIT = max(1, COMPILER_PARALLEL_LIMIT)
-COMPILER_PARALLEL_SEMAPHORE = threading.BoundedSemaphore(COMPILER_PARALLEL_LIMIT)
-
-DEFAULT_TAP_MAX_WORKERS = 8
-try:
-    TAP_MAX_WORKERS = int(
-        os.environ.get("KGPC_TAP_MAX_WORKERS", str(DEFAULT_TAP_MAX_WORKERS))
-    )
-except ValueError:
-    TAP_MAX_WORKERS = DEFAULT_TAP_MAX_WORKERS
-TAP_MAX_WORKERS = max(1, TAP_MAX_WORKERS)
-
-# Meson exposes toggleable behaviour via environment variables so CI can
-# selectively disable particularly slow checks such as the valgrind leak test.
-RUN_VALGRIND_TESTS = os.environ.get("RUN_VALGRIND_TESTS", "false").lower() in (
-    "1",
-    "true",
-    "yes",
+from tests.harness.env import (
+    EXE_EXT,
+    EXEC_TIMEOUT,
+    COMPILER_TIMEOUT,
+    LINK_TIMEOUT,
+    TEST_CASE_TIMEOUT,
+    PARALLEL_WORKERS,
+    TAP_MAX_WORKERS,
+    FPC_RTL_MODE,
+    FPC_RTL_DIR,
+    FPC_RTL_GENERATED_UNITS_DIRNAME,
+    PP_BOOTSTRAP_FULL_CHAIN_TIMEOUT,
+    IS_WINDOWS_ABI,
+    IS_WINE,
+    KGPC_PATH,
+    TEST_CASES_DIR,
+    INPUT_DATA_DIR,
+    TEST_OUTPUT_DIR,
+    GOLDEN_AST_DIR,
+    PROJECT_ROOT,
+    FAILURE_ARTIFACT_DIR,
+    RUN_VALGRIND_TESTS,
+    VALGRIND_MODE,
+    build_dir,
 )
 
-# Check if VALGRIND environment variable is set to enable valgrind for all tests
-VALGRIND_MODE = os.environ.get("VALGRIND", "false").lower() in ("1", "true", "yes")
+from tests.harness.cache import (
+    FPC_RTL_FLAGS,
+    with_fpc_rtl_ast_cache as _with_fpc_rtl_ast_cache,
+    test_cache_check as _test_cache_check,
+    test_cache_store as _test_cache_store,
+    _FPC_RTL_AST_CACHE_DIR,
+    _FPC_RTL_CODEGEN_CACHE_DIR,
+    _RUNTIME_LIB_PATH,
+)
 
-# Track how long individual compiler invocations take so we can identify the
-# slowest scenarios when the test suite finishes. The collected data is emitted
-# from TestCompiler.tearDownClass() and written to stderr to keep TAP output
-# intact.
-COMPILER_RUNS = []
-LINK_ARGS_BY_ASM = {}
+from tests.harness.sanitizers import (
+    HAS_PTY,
+    run_executable_with_valgrind,
+    _write_helper_script,
+    _run_helper_with_valgrind,
+)
 
-FAILURE_ARTIFACT_DIR_ENV = os.environ.get("KGPC_CI_FAILURE_DIR")
-FAILURE_ARTIFACT_DIR = Path(FAILURE_ARTIFACT_DIR_ENV) if FAILURE_ARTIFACT_DIR_ENV else None
+from tests.harness.runner import (
+    COMPILER_RUNS,
+    LINK_ARGS_BY_ASM,
+    run_compiler,
+    is_coverage_enabled,
+    EXPLICIT_TARGET_FLAGS,
+    _has_explicit_target_flag,
+)
 
-# FPC RTL test mode: compile test cases against the Free Pascal Compiler RTL
-# instead of KGPC's own runtime. Set KGPC_FPC_RTL=1 to enable.
-FPC_RTL_MODE = os.environ.get("KGPC_FPC_RTL", "").lower() in ("1", "true", "yes")
-FPC_RTL_DIR = os.path.join(os.environ.get("KGPC_FPC_RTL_DIR", "FPCSource"), "rtl")
+from tests.harness.reporters import (
+    TAPTestResult,
+    TAPTestRunner,
+    TimingTestResult,
+)
 
-# AST cache directory for FPC RTL mode — avoids re-parsing system.pp + objpas.pp
-# for every test (saves ~3.5s per test).
-# The cache is automatically invalidated when the compiler binary changes.
-_FPC_RTL_AST_CACHE_DIR = None
-if FPC_RTL_MODE:
-    _FPC_RTL_AST_CACHE_DIR = os.path.join(
-        os.environ.get("MESON_BUILD_ROOT", "builddir"), "fpc_rtl_ast_cache"
-    )
-    # KGPC validates AST cache entries against the compiler binary mtime.
-    # Keep a sentinel for diagnostics, but do not delete the cache directory:
-    # stale entries are rejected by the compiler and can coexist safely.
-    _cache_sentinel = os.path.join(_FPC_RTL_AST_CACHE_DIR, ".compiler_mtime")
-    _compiler_mtime = ""
-    if os.path.exists(KGPC_PATH):
-        st = os.stat(KGPC_PATH)
-        _compiler_mtime = f"{st.st_mtime_ns}:{st.st_size}"
-    os.makedirs(_FPC_RTL_AST_CACHE_DIR, exist_ok=True)
-    if _compiler_mtime:
-        with open(_cache_sentinel, "w") as f:
-            f.write(_compiler_mtime)
+from tests.harness.parallel import (
+    TAPParallelTestResult,
+    TAPParallelTestRunner,
+    ParallelTestResult,
+    ParallelTestRunner,
+    _flatten_tests,
+    _prepare_parallel_class_fixtures,
+    _cleanup_parallel_class_fixtures,
+    _run_single_test_with_timeout,
+)
 
-FPC_RTL_FLAGS = [
-    "--no-stdlib",
-    "-I" + os.path.join(FPC_RTL_DIR, "linux"),
-    "-I" + os.path.join(FPC_RTL_DIR, "linux", "x86_64"),
-    "-I" + os.path.join(FPC_RTL_DIR, "inc"),
-    "-I" + os.path.join(FPC_RTL_DIR, "unix"),
-    "-I" + os.path.join(FPC_RTL_DIR, "x86_64"),
-    "-I" + os.path.join(FPC_RTL_DIR, "objpas", "sysutils"),
-    "-I" + os.path.join(FPC_RTL_DIR, "objpas", "classes"),
-    "-Fu" + os.path.join(FPC_RTL_DIR, "objpas"),
-    "-Fu" + os.path.join(FPC_RTL_DIR, "inc"),
-    "-Fu" + os.path.join(FPC_RTL_DIR, "linux"),
-    "-Fu" + os.path.join(FPC_RTL_DIR, "linux", "x86_64"),
-    "-Fu" + os.path.join(FPC_RTL_DIR, "x86_64"),
-    "-Fu" + os.path.join(FPC_RTL_DIR, "unix"),
-    "-I" + os.path.join(os.environ.get("KGPC_FPC_RTL_DIR", "FPCSource"), "packages", "rtl-objpas", "src", "inc"),
-    "-Fu" + os.path.join(os.environ.get("KGPC_FPC_RTL_DIR", "FPCSource"), "packages", "rtl-objpas", "src", "inc"),
-    "-I" + os.path.join(os.environ.get("KGPC_FPC_RTL_DIR", "FPCSource"), "packages", "rtl-console", "src", "inc"),
-    "-Fu" + os.path.join(os.environ.get("KGPC_FPC_RTL_DIR", "FPCSource"), "packages", "rtl-console", "src", "inc"),
-    "-Fu" + os.path.join(os.environ.get("KGPC_FPC_RTL_DIR", "FPCSource"), "packages", "rtl-console", "src", "unix"),
-]
-if _FPC_RTL_AST_CACHE_DIR is not None:
-    FPC_RTL_FLAGS.append("--pp-cache-dir=" + _FPC_RTL_AST_CACHE_DIR)
+from tests.harness.artifacts import (
+    read_file_content,
+    store_failure_artifacts as _store_failure_artifacts,
+    _signal_name_suffix,
+)
 
-PP_BOOTSTRAP_FULL_CHAIN_TIMEOUT = 1800
-FPC_RTL_GENERATED_UNITS_DIRNAME = "units"
+from tests.harness.discovery import (
+    kgpc_bootstrap_flags as _kgpc_bootstrap_flags_impl,
+    pp_bootstrap_compiler_flags as _pp_bootstrap_compiler_flags,
+    pp_bootstrap_program_flags as _pp_bootstrap_program_flags,
+    should_include_in_fpcrtl as _should_include_in_fpcrtl_impl,
+    _strip_pascal_comments,
+    _extract_used_units,
+    _collect_pascal_unit_names,
+    _get_fpc_rtl_known_units,
+    _tree_contains_newer_file as _tree_contains_newer_file_impl,
+)
 
-
-def _with_fpc_rtl_ast_cache(flags):
-    flags = list(flags)
-    if _FPC_RTL_AST_CACHE_DIR is not None:
-        flags.append("--pp-cache-dir=" + _FPC_RTL_AST_CACHE_DIR)
-    return flags
-
-
-def _bootstrap_define_flags():
-    """Returns KGPC-style define flags for invoking the KGPC compiler itself."""
-    return [
-        "-DCPU64",
-        "-DCPUX86_64",
-        "-Dx86_64",
-        "-DFPC",
-        "-DLINUX",
-        "-DUNIX",
-        "-DFPC_HAS_TYPE_EXTENDED",
-        "-DSUPPORT_EXTENDED",
-        "-DFPC_BOOTSTRAP_INDIRECT_ENTRY",
-        "-Sg",
-    ]
-
-
-def _fpc_bootstrap_define_flags():
-    """Returns FPC-style define flags for invoking a compiled FPC binary (pp_bootstrap etc.)."""
-    return [
-        "-dCPU64",
-        "-dCPUX86_64",
-        "-dx86_64",
-        "-dFPC",
-        "-dLINUX",
-        "-dUNIX",
-        "-dFPC_HAS_TYPE_EXTENDED",
-        "-dSUPPORT_EXTENDED",
-        "-dFPC_BOOTSTRAP_INDIRECT_ENTRY",
-        "-Sg",
-    ]
-
-
-def _bootstrap_rtl_include_dirs(fpc_src):
-    return [
-        os.path.join(fpc_src, "rtl", "objpas"),
-        os.path.join(fpc_src, "rtl", "objpas", "sysutils"),
-        os.path.join(fpc_src, "rtl", "objpas", "classes"),
-        os.path.join(fpc_src, "rtl", "linux"),
-        os.path.join(fpc_src, "rtl", "unix"),
-        os.path.join(fpc_src, "rtl", "inc"),
-        os.path.join(fpc_src, "rtl", "x86_64"),
-        os.path.join(fpc_src, "rtl", "linux", "x86_64"),
-        os.path.join(fpc_src, "rtl", "unix", "x86_64"),
-    ]
-
-
-def _bootstrap_rtl_unit_dirs(fpc_src):
-    return [
-        os.path.join(fpc_src, "rtl", "unix"),
-        os.path.join(fpc_src, "rtl", "linux"),
-        os.path.join(fpc_src, "rtl", "objpas"),
-        os.path.join(fpc_src, "rtl", "inc"),
-        os.path.join(fpc_src, "rtl", "objpas", "sysutils"),
-        os.path.join(fpc_src, "rtl", "objpas", "classes"),
-        # FPC charmap units (cp1250.pas, cp1251.pas, …, cp8859_1.pas) each
-        # declare a same-named `unicodemap` typed-const that registers itself
-        # via `registermapping(@unicodemap)` in initialization.  pp_bootstrap
-        # consults this chain via getmap(); if KGPC's per-unit storage keys
-        # are wrong the chain becomes self-referential and runtime lookups
-        # (e.g. widestr.pas:391) crash.
-        os.path.join(fpc_src, "rtl", "charmaps"),
-    ]
-
-
-def _bootstrap_compiler_include_dirs(fpc_src):
-    return [
-        os.path.join(fpc_src, "compiler"),
-        os.path.join(fpc_src, "compiler", "x86"),
-        os.path.join(fpc_src, "compiler", "x86_64"),
-    ]
-
-
-def _bootstrap_compiler_unit_dirs(fpc_src):
-    return [
-        os.path.join(fpc_src, "compiler"),
-        os.path.join(fpc_src, "compiler", "x86"),
-        os.path.join(fpc_src, "compiler", "x86_64"),
-        os.path.join(fpc_src, "compiler", "systems"),
-    ]
+# ---------------------------------------------------------------------------
+# Local wrappers that preserve the original call signatures used in the
+# test class and discovery functions below.
+# ---------------------------------------------------------------------------
 
 
 def _kgpc_bootstrap_flags(fpc_src, *, include_compiler_dirs):
-    flags = ["--no-stdlib"]
-    flags.extend(_bootstrap_define_flags())
-    flags.extend("-I" + path for path in _bootstrap_rtl_include_dirs(fpc_src))
-    if include_compiler_dirs:
-        flags.extend("-I" + path for path in _bootstrap_compiler_include_dirs(fpc_src))
-    flags.extend("-Fu" + path for path in _bootstrap_rtl_unit_dirs(fpc_src))
-    if include_compiler_dirs:
-        flags.extend("-Fu" + path for path in _bootstrap_compiler_unit_dirs(fpc_src))
-    flags = _with_fpc_rtl_ast_cache(flags)
-    if _FPC_RTL_CODEGEN_CACHE_DIR is not None:
-        flags.append("--codegen-cache-dir=" + _FPC_RTL_CODEGEN_CACHE_DIR)
-    return flags
-
-
-def _pp_bootstrap_compiler_flags(
-    fpc_src, *, rtl_units_dir, output_dir, unit_output_dir, executable_name
-):
-    absolute_fpc_src = os.path.abspath(fpc_src)
-    flags = [
-        "-n",
-        "-FE" + output_dir,
-        "-FU" + unit_output_dir,
-        "-o" + executable_name,
-        "-Fu" + rtl_units_dir,
-    ]
-    flags.extend(_fpc_bootstrap_define_flags())
-    # pp_bootstrap rebuilds pp.pas with cwd inside FPCSource/compiler, so these
-    # paths must be absolute rather than repository-root-relative.
-    flags.extend(
-        "-Fi" + path for path in _bootstrap_compiler_include_dirs(absolute_fpc_src)
+    """Wrapper preserving original signature; injects cache dirs from cache module."""
+    return _kgpc_bootstrap_flags_impl(
+        fpc_src,
+        include_compiler_dirs=include_compiler_dirs,
+        fpc_rtl_ast_cache_dir=_FPC_RTL_AST_CACHE_DIR,
+        fpc_rtl_codegen_cache_dir=_FPC_RTL_CODEGEN_CACHE_DIR,
     )
-    flags.extend(
-        "-Fu" + path for path in _bootstrap_compiler_unit_dirs(absolute_fpc_src)
-    )
-    return flags
-
-
-def _pp_bootstrap_program_flags(*, rtl_units_dir, output_dir, executable_name):
-    return [
-        "-n",
-        "-FE" + output_dir,
-        "-o" + executable_name,
-        "-Fu" + rtl_units_dir,
-    ]
 
 
 def _tree_contains_newer_file(root_dir, reference_file):
-    if not os.path.isdir(root_dir) or not os.path.isfile(reference_file):
-        return False
-    reference_mtime = os.path.getmtime(reference_file)
-    for current_root, dirnames, filenames in os.walk(root_dir):
-        # Skip generated RTL output so only source-tree changes invalidate the
-        # same-source unit cache. This must rewrite dirnames in place so
-        # os.walk does not descend into the filtered directories.
-        dirnames[:] = [
-            dirname for dirname in dirnames
-            if dirname != FPC_RTL_GENERATED_UNITS_DIRNAME
-        ]
-        for filename in filenames:
-            path = os.path.join(current_root, filename)
-            if os.path.getmtime(path) > reference_mtime:
-                return True
-    return False
-
-
-def _signal_name_suffix(returncode):
-    if returncode >= 0:
-        return ""
-    try:
-        return f" ({signal.Signals(-returncode).name})"
-    except (ValueError, AttributeError):
-        return ""
-
-# Codegen object cache: the compiler assembles full .o on cache miss,
-# then uses --skip-unit-codegen on cache hit (all handled internally).
-_FPC_RTL_CODEGEN_CACHE_DIR = None
-if FPC_RTL_MODE:
-    _FPC_RTL_CODEGEN_CACHE_DIR = os.path.join(
-        os.environ.get("MESON_BUILD_ROOT", "builddir"), "fpc_rtl_codegen_cache"
+    """Wrapper preserving original 2-arg signature."""
+    return _tree_contains_newer_file_impl(
+        root_dir, reference_file, FPC_RTL_GENERATED_UNITS_DIRNAME
     )
-    # The compiler includes its own mtime in the cache key, so old entries
-    # are never matched after a rebuild. Do not delete the directory here:
-    # cache invalidation belongs in the key/freshness checks, not in test
-    # harness cleanup.
-    _cg_sentinel = os.path.join(
-        os.environ.get("MESON_BUILD_ROOT", "builddir"), ".codegen_cache_mtime"
+
+
+def _should_include_in_fpcrtl(base_name, pascal_file):
+    """Wrapper preserving original signature; passes FPC_RTL_IMPLICIT_UNIT_TESTS."""
+    return _should_include_in_fpcrtl_impl(
+        base_name, pascal_file, FPC_RTL_IMPLICIT_UNIT_TESTS
     )
-    os.makedirs(_FPC_RTL_CODEGEN_CACHE_DIR, exist_ok=True)
-    if _compiler_mtime:
-        with open(_cg_sentinel, "w") as f:
-            f.write(_compiler_mtime)
-    FPC_RTL_FLAGS.append("--codegen-cache-dir=" + _FPC_RTL_CODEGEN_CACHE_DIR)
+
 
 # ---------------------------------------------------------------------------
-# Background AST cache primer for FPC RTL mode.
-# Spawns a lightweight compiler process that compiles a minimal Pascal program
-# importing the most common RTL units (SysUtils, Classes).  This populates the
-# AST cache so that subsequent parallel test compilations get cache hits
-# instead of all redundantly parsing system.pp.
-#
-# The primer runs outside the compiler semaphore — it does NOT block test
-# execution.  It runs concurrently with the first batch of tests.  Tests
-# that start before the primer finishes will parse from source (the AST
-# cache uses atomic write-then-rename, so concurrent writes are safe).
-# Tests that start after the primer finishes benefit from cached ASTs.
+# Constants that were in the original monolith between line ~530 and ~645
+# and are referenced by TestCompiler / discovery functions below.
 # ---------------------------------------------------------------------------
-_CACHE_PRIMER_PROC = None
-if FPC_RTL_MODE and _FPC_RTL_AST_CACHE_DIR is not None:
-    # Only prime if the AST cache is empty (cold start).
-    _existing_cache_files = [
-        f for f in os.listdir(_FPC_RTL_AST_CACHE_DIR)
-        if f.endswith(".ast_cache")
-    ]
-    if not _existing_cache_files:
-        import tempfile as _tempfile
-        _primer_src = os.path.join(
-            _tempfile.gettempdir(), "kgpc_cache_primer.p"
-        )
-        _primer_asm = os.path.join(
-            _tempfile.gettempdir(), "kgpc_cache_primer.s"
-        )
-        with open(_primer_src, "w") as _f:
-            _f.write("program CachePrimer;\nuses SysUtils, Classes;\nbegin\nend.\n")
-        _primer_cmd = [KGPC_PATH, _primer_src, _primer_asm, "--emit-link-args"] + FPC_RTL_FLAGS
-        print(f"--- Launching background AST cache primer ---", file=sys.stderr)
-        sys.stderr.flush()
-        _CACHE_PRIMER_PROC = subprocess.Popen(
-            _primer_cmd,
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-        )
-    del _existing_cache_files
-# Test result cache — skip compile/link/run when nothing changed.
-# Cache key = hash of: compiler binary, test .p file, .expected file,
-# flags, runtime library.  Stored as empty files named by the key hash
-# in _TEST_RESULT_CACHE_DIR.
-# ---------------------------------------------------------------------------
-_TEST_RESULT_CACHE_DIR = os.path.join(
-    os.environ.get("MESON_BUILD_ROOT", "build"), "test_result_cache"
-)
-os.makedirs(_TEST_RESULT_CACHE_DIR, exist_ok=True)
-
-# Pre-compute compiler identity once (mtime + size — fast, changes on every rebuild).
-_COMPILER_HASH = ""
-if os.path.exists(KGPC_PATH):
-    _st = os.stat(KGPC_PATH)
-    _COMPILER_HASH = f"{_st.st_mtime_ns}:{_st.st_size}"
-
-# Pre-compute runtime library identity once.
-_RUNTIME_LIB_PATH = os.environ.get("KGPC_RUNTIME_LIB", "")
-_RUNTIME_HASH = ""
-if _RUNTIME_LIB_PATH and os.path.exists(_RUNTIME_LIB_PATH):
-    _st = os.stat(_RUNTIME_LIB_PATH)
-    _RUNTIME_HASH = f"{_st.st_mtime_ns}:{_st.st_size}"
-
-
-def _file_hash(path):
-    """Return mtime:size identity string for a file, or empty string if missing."""
-    if not path or not os.path.exists(path):
-        return ""
-    st = os.stat(path)
-    return f"{st.st_mtime_ns}:{st.st_size}"
-
-
-def _test_cache_key(input_file, expected_file, flags):
-    """Compute a cache key for a test case."""
-    h = hashlib.sha256()
-    h.update(_COMPILER_HASH.encode())
-    h.update(_RUNTIME_HASH.encode())
-    h.update(_file_hash(input_file).encode())
-    h.update(_file_hash(expected_file).encode())
-    # Include hashes of any auxiliary unit files (e.g. cross_unit_inherit_base.p)
-    base_name = os.path.splitext(os.path.basename(input_file))[0]
-    test_dir = os.path.dirname(input_file)
-    for f in sorted(os.listdir(test_dir)):
-        if f.startswith(base_name + "_") and f.endswith(".p"):
-            h.update(_file_hash(os.path.join(test_dir, f)).encode())
-    # Include .input file if present
-    input_data = os.path.join(test_dir, base_name + ".input")
-    h.update(_file_hash(input_data).encode())
-    h.update(repr(sorted(flags)).encode() if flags else b"")
-    return h.hexdigest()
-
-
-def _test_cache_check(input_file, expected_file, flags):
-    """Return True if this test has a cached passing result."""
-    key = _test_cache_key(input_file, expected_file, flags)
-    return os.path.exists(os.path.join(_TEST_RESULT_CACHE_DIR, key))
-
-
-def _test_cache_store(input_file, expected_file, flags):
-    """Record a passing result for this test."""
-    key = _test_cache_key(input_file, expected_file, flags)
-    path = os.path.join(_TEST_RESULT_CACHE_DIR, key)
-    with open(path, "w") as f:
-        f.write("")
-
 
 UNIT_ONLY_TESTS = {
     "directives_and_properties_unit",
@@ -552,1027 +181,6 @@ UNIT_ONLY_TESTS = {
 UNIT_ONLY_FLAGS = {
     "tdd_variant_shadow_record": ["--no-stdlib"],
 }
-
-
-def _sanitize_test_identifier(name):
-    return (
-        name.replace("/", "_")
-        .replace("\\", "_")
-        .replace(":", "_")
-        .replace(" ", "_")
-    )
-
-
-def _copy_artifact(path, dest_dir):
-    if not path:
-        return
-    src = Path(path)
-    if not src.exists():
-        return
-    try:
-        shutil.copy2(src, dest_dir / src.name)
-    except OSError:
-        pass
-
-
-def _write_artifact_text(dest_dir, filename, content):
-    if content is None:
-        return
-    target = dest_dir / filename
-    target.write_text(str(content), encoding="utf-8", errors="ignore")
-
-
-def _store_failure_artifacts(
-    test_id,
-    base_name=None,
-    *,
-    input_file=None,
-    asm_file=None,
-    executable_file=None,
-    expected_file=None,
-    compiler_output=None,
-    normalized_output=None,
-    raw_stdout=None,
-    raw_stderr=None,
-    expected_output=None,
-    expected_stderr=None,
-    returncode=None,
-    exception_text=None,
-):
-    if FAILURE_ARTIFACT_DIR is None:
-        return
-
-    case_name = base_name or test_id.split(".")[-1]
-    dest = FAILURE_ARTIFACT_DIR / _sanitize_test_identifier(case_name)
-    dest.mkdir(parents=True, exist_ok=True)
-
-    info_lines = [
-        f"test_id={test_id}",
-        f"case={case_name}",
-    ]
-    if returncode is not None:
-        info_lines.append(f"returncode={returncode}")
-    (dest / "info.txt").write_text(
-        "\n".join(info_lines), encoding="utf-8", errors="ignore"
-    )
-
-    for candidate in (input_file, asm_file, executable_file, expected_file):
-        _copy_artifact(candidate, dest)
-
-    _write_artifact_text(dest, "compiler-stderr.txt", compiler_output)
-    _write_artifact_text(dest, "raw-stdout.txt", raw_stdout)
-    _write_artifact_text(dest, "raw-stderr.txt", raw_stderr)
-    _write_artifact_text(dest, "normalized-output.txt", normalized_output)
-    _write_artifact_text(dest, "expected-output.txt", expected_output)
-    _write_artifact_text(dest, "expected-stderr.txt", expected_stderr)
-    if exception_text:
-        _write_artifact_text(dest, "exception.txt", exception_text)
-
-# Flags that explicitly request a target ABI so we do not override them.
-EXPLICIT_TARGET_FLAGS = {
-    "--target",
-    "-target",
-    "--target-windows",
-    "-target-windows",
-    "--windows-abi",
-    "--target-sysv",
-    "-target-sysv",
-    "--sysv-abi",
-}
-
-_COVERAGE_ENABLED_CACHE = None
-
-
-def is_coverage_enabled():
-    global _COVERAGE_ENABLED_CACHE
-    if _COVERAGE_ENABLED_CACHE is not None:
-        return _COVERAGE_ENABLED_CACHE
-
-    build_dir = os.environ.get("MESON_BUILD_ROOT", "build")
-    options_path = os.path.join(build_dir, "meson-info", "intro-buildoptions.json")
-    try:
-        with open(options_path, "r", encoding="utf-8") as f:
-            data = json.load(f)
-        _COVERAGE_ENABLED_CACHE = next(
-            (
-                bool(opt.get("value"))
-                for opt in data
-                if opt.get("name") == "b_coverage"
-            ),
-            False,
-        )
-    except (OSError, json.JSONDecodeError):
-        _COVERAGE_ENABLED_CACHE = False
-    return _COVERAGE_ENABLED_CACHE
-
-
-def _has_explicit_target_flag(flags):
-    if not flags:
-        return False
-    for flag in flags:
-        if flag in EXPLICIT_TARGET_FLAGS:
-            return True
-        if flag.startswith("--target=") or flag.startswith("-target="):
-            return True
-    return False
-
-# The compiler is built by Meson now, so this function is not needed.
-
-
-def run_executable_with_valgrind(executable_args, **kwargs):
-    """Run an executable, optionally with valgrind in CI mode."""
-    command = list(executable_args)
-    explicit_pty_capture = kwargs.pop("use_pty_capture", None)
-    
-    # Use valgrind when CI mode is enabled and valgrind is available
-    if VALGRIND_MODE and shutil.which("valgrind") is not None:
-        valgrind_cmd = [
-            "valgrind",
-            "--tool=memcheck",
-            "--track-origins=yes",
-            "--num-callers=50",
-            "--error-exitcode=1",
-        ]
-        command = valgrind_cmd + command
-        print(f"--- Running executable with valgrind: {' '.join(command)} ---", file=sys.stderr)
-
-        if "timeout" not in kwargs:
-            kwargs["timeout"] = EXEC_TIMEOUT
-        return subprocess.run(command, **kwargs)
-
-    if IS_WINE:
-        text_mode = bool(kwargs.get("text"))
-        input_data = kwargs.get("input")
-        check = bool(kwargs.get("check"))
-        capture_output = bool(kwargs.get("capture_output"))
-        timeout = kwargs.get("timeout", EXEC_TIMEOUT)
-
-        if timeout is not None:
-            # Wine's Windows-Python subprocess waits can fail with WinError 6
-            # for short-lived child processes. Use shell redirection instead.
-            timeout = None
-
-        with tempfile.TemporaryDirectory() as tmpdir:
-            stdout_path = os.path.join(tmpdir, "stdout.txt")
-            stderr_path = os.path.join(tmpdir, "stderr.txt")
-            stdin_path = os.path.join(tmpdir, "stdin.txt")
-
-            if input_data is not None:
-                mode = "w" if text_mode else "wb"
-                with open(stdin_path, mode) as handle:
-                    handle.write(input_data)
-
-            command_line = subprocess.list2cmdline(command)
-            redirect_bits = []
-            if input_data is not None:
-                redirect_bits.append(f'< "{stdin_path}"')
-            redirect_bits.append(f'> "{stdout_path}"')
-            redirect_bits.append(f'2> "{stderr_path}"')
-            shell_command = f"{command_line} {' '.join(redirect_bits)}"
-
-            returncode = os.system(shell_command)
-
-            stdout_data = None
-            stderr_data = None
-            if capture_output:
-                read_mode = "r" if text_mode else "rb"
-                read_kwargs = (
-                    {
-                        "encoding": locale.getpreferredencoding(False) or "cp1252",
-                        "errors": "replace",
-                    }
-                    if text_mode
-                    else {}
-                )
-                with open(stdout_path, read_mode, **read_kwargs) as handle:
-                    stdout_data = handle.read()
-                with open(stderr_path, read_mode, **read_kwargs) as handle:
-                    stderr_data = handle.read()
-
-            completed = subprocess.CompletedProcess(
-                args=command,
-                returncode=returncode,
-                stdout=stdout_data,
-                stderr=stderr_data,
-            )
-            if check and returncode != 0:
-                raise subprocess.CalledProcessError(
-                    returncode, command, output=completed.stdout, stderr=completed.stderr
-                )
-            return completed
-
-    # PTY capture is opt-in. Most tests, and validation against real FPC invoked via
-    # subprocess pipes, should observe redirected-output behaviour rather than a fake TTY.
-    # Enable KGPC_USE_PTY_CAPTURE=1 when a test genuinely needs terminal semantics.
-    #
-    # Important: keep stdin as a pipe when input=... is provided, so test input is
-    # not echoed by terminal line discipline.
-    if explicit_pty_capture is None:
-        use_pty_capture = os.environ.get("KGPC_USE_PTY_CAPTURE", "").lower() in ("1", "true", "yes")
-    else:
-        use_pty_capture = bool(explicit_pty_capture)
-    if use_pty_capture and HAS_PTY and kwargs.get("capture_output") and "stdout" not in kwargs and "stderr" not in kwargs:
-        text_mode = bool(kwargs.get("text"))
-        timeout = kwargs.get("timeout", EXEC_TIMEOUT)
-        input_data = kwargs.get("input")
-        check = bool(kwargs.get("check"))
-
-        master_fd, slave_fd = pty.openpty()
-        try:
-            try:
-                os.set_blocking(master_fd, False)
-            except AttributeError:
-                flags = fcntl.fcntl(master_fd, fcntl.F_GETFL)
-                fcntl.fcntl(master_fd, fcntl.F_SETFL, flags | os.O_NONBLOCK)
-            attrs = termios.tcgetattr(slave_fd)
-            attrs[3] &= ~termios.ECHO
-            termios.tcsetattr(slave_fd, termios.TCSANOW, attrs)
-
-            proc = subprocess.Popen(
-                command,
-                stdin=subprocess.PIPE if input_data is not None else None,
-                stdout=slave_fd,
-                stderr=slave_fd,
-                text=text_mode,
-                close_fds=True,
-            )
-
-            os.close(slave_fd)
-            slave_fd = None
-
-            if input_data is not None:
-                proc.stdin.write(input_data)
-                proc.stdin.close()
-
-            output_bytes = bytearray()
-            start = time.monotonic()
-            while True:
-                if timeout is not None and (time.monotonic() - start) > timeout:
-                    proc.kill()
-                    break
-
-                r, _, _ = select.select([master_fd], [], [], 0.1)
-                if r:
-                    try:
-                        chunk = os.read(master_fd, 4096)
-                    except BlockingIOError:
-                        chunk = b""
-                    except OSError:
-                        break
-                    if chunk:
-                        output_bytes.extend(chunk)
-
-                if proc.poll() is not None and not r:
-                    break
-
-            # Drain remaining bytes
-            while True:
-                try:
-                    chunk = os.read(master_fd, 4096)
-                except BlockingIOError:
-                    break
-                except OSError:
-                    break
-                if not chunk:
-                    break
-                output_bytes.extend(chunk)
-
-            returncode = proc.wait()
-
-            if text_mode:
-                stdout_text = output_bytes.decode("utf-8", errors="replace")
-                while "\r\r\n" in stdout_text:
-                    stdout_text = stdout_text.replace("\r\r\n", "\r\n")
-                stdout_text = stdout_text.replace("\r\n", "\n").replace("\r", "")
-                stdout_val = stdout_text
-                stderr_val = ""
-            else:
-                stdout_val = bytes(output_bytes)
-                stderr_val = b""
-
-            completed = subprocess.CompletedProcess(
-                args=command,
-                returncode=returncode,
-                stdout=stdout_val,
-                stderr=stderr_val,
-            )
-            if check and returncode != 0:
-                raise subprocess.CalledProcessError(
-                    returncode, command, output=completed.stdout, stderr=completed.stderr
-                )
-            return completed
-        finally:
-            if slave_fd is not None:
-                os.close(slave_fd)
-            os.close(master_fd)
-
-    if "timeout" not in kwargs:
-        kwargs["timeout"] = EXEC_TIMEOUT
-    return subprocess.run(command, **kwargs)
-
-
-def _write_helper_script(tmpdir, body):
-    """Write a small Python helper script used to exercise PTY/non-PTY paths."""
-    script = os.path.join(tmpdir, "pty_helper.py")
-    with open(script, "w", encoding="utf-8") as handle:
-        handle.write(
-            textwrap.dedent(
-                f"""\
-                #!/usr/bin/env {sys.executable}
-                import sys
-                import time
-
-                def main():
-            """
-            )
-        )
-        for line in textwrap.dedent(body).splitlines():
-            handle.write(f"    {line}\n")
-        handle.write(
-            textwrap.dedent(
-                """\
-                if __name__ == "__main__":
-                    main()
-                """
-            )
-        )
-    os.chmod(script, 0o755)
-    return script
-
-
-def _run_helper_with_valgrind(command, *, timeout=None, text=True, input_data=None,
-                              capture_output=True, check=False, **kwargs):
-    """Wrapper to keep PTY helper tests consistent."""
-    runner_kwargs = {
-        "timeout": timeout if timeout is not None else EXEC_TIMEOUT,
-        "text": text,
-        "capture_output": capture_output,
-        "check": check,
-    }
-    if input_data is not None:
-        runner_kwargs["input"] = input_data
-    runner_kwargs.update(kwargs)
-    return run_executable_with_valgrind(command, **runner_kwargs)
-
-
-def run_compiler(input_file, output_file, flags=None, timeout=None):
-    """Runs the KGPC compiler with the given arguments."""
-    if flags is None:
-        flags = []
-    else:
-        flags = list(flags)
-    if timeout is None:
-        timeout = COMPILER_TIMEOUT
-
-    # Ensure the output directory exists
-    os.makedirs(os.path.dirname(output_file), exist_ok=True)
-
-    command = [KGPC_PATH, input_file, output_file]
-    if "--emit-link-args" not in flags:
-        command.append("--emit-link-args")
-    if IS_WINDOWS_ABI and not _has_explicit_target_flag(flags):
-        command.append("--target-windows")
-    command.extend(flags)
-    
-    # Use valgrind when VALGRIND mode is enabled
-    if VALGRIND_MODE and shutil.which("valgrind") is not None:
-        valgrind_cmd = [
-            "valgrind",
-            "--tool=memcheck",
-            "--track-origins=yes",
-            "--num-callers=50",
-            "--error-exitcode=1",
-        ]
-        command = valgrind_cmd + command
-        print(f"--- Running compiler with valgrind: {' '.join(command)} ---", file=sys.stderr)
-        sys.stderr.flush()
-    else:
-        print(f"--- Running compiler: {' '.join(command)} ---", file=sys.stderr)
-        sys.stderr.flush()
-
-    start = time.perf_counter()
-    try:
-        run_kwargs = {
-            "check": True,
-            "capture_output": True,
-            "text": True,
-            "timeout": timeout,
-        }
-        with COMPILER_PARALLEL_SEMAPHORE:
-            result = subprocess.run(command, **run_kwargs)
-        duration = time.perf_counter() - start
-        for line in result.stderr.splitlines():
-            if line.startswith("KGPC_LINK_ARGS:"):
-                raw_args = line[len("KGPC_LINK_ARGS:"):].strip()
-                LINK_ARGS_BY_ASM[output_file] = raw_args.split() if raw_args else []
-        COMPILER_RUNS.append(
-            {
-                "command": command,
-                "duration": duration,
-                "returncode": result.returncode,
-            }
-        )
-        print(
-            result.stderr, file=sys.stderr
-        )  # The compiler prints status messages to stderr
-        return result.stderr
-    except subprocess.CalledProcessError as e:
-        duration = time.perf_counter() - start
-        COMPILER_RUNS.append(
-            {
-                "command": command,
-                "duration": duration,
-                "returncode": e.returncode,
-            }
-        )
-        print(f"--- Compiler execution failed ---", file=sys.stderr)
-        print(f"--- stdout: {e.stdout} ---", file=sys.stderr)
-        print(f"--- stderr: {e.stderr} ---", file=sys.stderr)
-        raise
-
-
-class TAPTestResult(unittest.TestResult):
-    # Enable verbose logging to stderr (controlled by env var for CI debugging)
-    VERBOSE_LOG = os.environ.get("KGPC_TEST_VERBOSE", "false").lower() in ("1", "true", "yes")
-
-    def __init__(self, stream):
-        super().__init__()
-        self.stream = stream
-        self._test_index = 0
-        self._test_states = {}
-        self._test_start_times = {}
-
-    def _emit(self, line):
-        text = f"{line}\n"
-        encoding = getattr(self.stream, "encoding", None) or "utf-8"
-        safe_text = text.encode(
-            encoding, errors="backslashreplace"
-        ).decode(encoding, errors="replace")
-        self.stream.write(safe_text)
-        self.stream.flush()
-
-    def _emit_diagnostic(self, text):
-        for raw_line in text.rstrip().splitlines():
-            self._emit(f"# {raw_line}")
-
-    def _test_name(self, test):
-        try:
-            return test.id()
-        except AttributeError:
-            return str(test)
-
-    def _subtest_name(self, subtest):
-        if subtest is None:
-            return ""
-        params = getattr(subtest, "params", None) or {}
-        msg = getattr(subtest, "message", None)
-        parts = []
-        if params:
-            formatted = ", ".join(
-                f"{key}={value!r}" for key, value in sorted(params.items())
-            )
-            parts.append(formatted)
-        if msg:
-            parts.append(str(msg))
-        if parts:
-            return " | ".join(parts)
-        return str(subtest)
-
-    def _get_state(self, test):
-        return self._test_states.setdefault(
-            test, {"reported": False, "had_failure": False}
-        )
-
-    def _emit_failure_header(self, test):
-        state = self._get_state(test)
-        if not state["reported"]:
-            self._emit(f"not ok {self._test_index} - {self._test_name(test)}")
-            state["reported"] = True
-        return state
-
-    def _mark_failure(self, test):
-        state = self._get_state(test)
-        state["had_failure"] = True
-        return state
-
-    def startTest(self, test):
-        super().startTest(test)
-        # Enforce per-test timeout in TAP mode to avoid hangs.
-        if not IS_WINDOWS_ABI and hasattr(signal, "SIGALRM"):
-            method = getattr(test, test._testMethodName, None)
-            per_test_timeout = getattr(method, '_timeout', TEST_CASE_TIMEOUT)
-
-            def _timeout_handler(_signum, _frame):
-                raise TimeoutError(f"Test timed out after {per_test_timeout} seconds")
-
-            self._prev_alarm_handler = signal.getsignal(signal.SIGALRM)
-            signal.signal(signal.SIGALRM, _timeout_handler)
-            signal.alarm(per_test_timeout)
-        self._test_index += 1
-        self._test_states[test] = {"reported": False, "had_failure": False}
-        self._test_start_times[test] = time.monotonic()
-        # Always emit starting message so we can see which test is running
-        self._emit(f"# [STARTING] {self._test_name(test)}")
-
-    def stopTest(self, test):
-        elapsed = 0.0
-        start_time = self._test_start_times.pop(test, None)
-        if start_time is not None:
-            elapsed = time.monotonic() - start_time
-        if not IS_WINDOWS_ABI and hasattr(signal, "SIGALRM"):
-            signal.alarm(0)
-            prev = getattr(self, "_prev_alarm_handler", None)
-            if prev is not None:
-                signal.signal(signal.SIGALRM, prev)
-        # Note: stopTest is called after addSuccess/addFailure, so the result line is already emitted
-        super().stopTest(test)
-        self._test_states.pop(test, None)
-
-    def addSuccess(self, test):
-        super().addSuccess(test)
-        state = self._test_states.get(test)
-        if state and state.get("had_failure"):
-            return
-        self._emit(f"ok {self._test_index} - {self._test_name(test)}")
-        if state:
-            state["reported"] = True
-
-    def addSkip(self, test, reason):
-        super().addSkip(test, reason)
-        self._emit(f"ok {self._test_index} - {self._test_name(test)} # SKIP {reason}")
-        state = self._test_states.get(test)
-        if state:
-            state["reported"] = True
-
-    def addExpectedFailure(self, test, err):
-        super().addExpectedFailure(test, err)
-        self._emit(
-            f"ok {self._test_index} - {self._test_name(test)} # TODO expected failure"
-        )
-        state = self._test_states.get(test)
-        if state:
-            state["reported"] = True
-
-    def addUnexpectedSuccess(self, test):
-        super().addUnexpectedSuccess(test)
-        self._emit(
-            f"not ok {self._test_index} - {self._test_name(test)} # Unexpected success"
-        )
-        self._mark_failure(test)
-
-    def addFailure(self, test, err):
-        super().addFailure(test, err)
-        self._emit_failure_header(test)
-        self._emit_diagnostic("Failure:")
-        self._emit_diagnostic("".join(traceback.format_exception(*err)))
-        self._mark_failure(test)
-
-    def addError(self, test, err):
-        super().addError(test, err)
-        self._emit_failure_header(test)
-        self._emit_diagnostic("Error:")
-        self._emit_diagnostic("".join(traceback.format_exception(*err)))
-        self._mark_failure(test)
-
-    def addSubTest(self, test, subtest, err):
-        super().addSubTest(test, subtest, err)
-        if err is None:
-            return
-        self._emit_failure_header(test)
-        description = self._subtest_name(subtest)
-        if description:
-            self._emit_diagnostic(f"Subtest failed: {description}")
-        else:
-            self._emit_diagnostic("Subtest failed")
-        self._emit_diagnostic("".join(traceback.format_exception(*err)))
-        self._mark_failure(test)
-
-
-class TAPTestRunner:
-    def __init__(self, stream=None, failfast=False):
-        self.stream = stream or sys.stdout
-        self.failfast = failfast
-
-    def run(self, test):
-        result = TAPTestResult(self.stream)
-        test_count = test.countTestCases()
-        result.startTestRun()
-        try:
-            result.stream.write(f"1..{test_count}\n")
-            result.stream.flush()
-            test(result)
-        finally:
-            result.stopTestRun()
-        return result
-
-
-def _flatten_tests(test):
-    """Recursively flatten a TestSuite into individual test cases."""
-    if isinstance(test, unittest.TestSuite):
-        for t in test:
-            yield from _flatten_tests(t)
-    else:
-        yield test
-
-
-def _prepare_parallel_class_fixtures(tests, log_stream=None):
-    """Run setUpClass once per class; return (class_errors, setup_ok_classes)."""
-    class_order = []
-    seen = set()
-    for test in tests:
-        cls = test.__class__
-        if cls not in seen:
-            seen.add(cls)
-            class_order.append(cls)
-
-    class_errors = {}
-    setup_ok_classes = []
-    for cls in class_order:
-        try:
-            cls.setUpClass()
-            setup_ok_classes.append(cls)
-        except Exception:
-            class_errors[cls] = traceback.format_exc()
-            if log_stream is not None:
-                log_stream.write(f"[PARALLEL] setUpClass failed for {cls.__name__}\n")
-                log_stream.flush()
-    return class_errors, setup_ok_classes
-
-
-def _cleanup_parallel_class_fixtures(setup_ok_classes, log_stream=None):
-    """Run tearDownClass for classes whose setUpClass succeeded."""
-    for cls in reversed(setup_ok_classes):
-        try:
-            cls.tearDownClass()
-        except Exception:
-            if log_stream is not None:
-                log_stream.write(f"[PARALLEL] tearDownClass failed for {cls.__name__}\n")
-                log_stream.flush()
-
-
-def _run_single_test_with_timeout(test, timeout, log_stream=None):
-    """Run a single test with a timeout. Returns (test, result_type, err_info)."""
-    test_id = test.id()
-    start_time = time.monotonic()
-    if log_stream is not None:
-        log_stream.write(f"[PARALLEL] Starting: {test_id}\n")
-        log_stream.flush()
-
-    result = unittest.TestResult()
-    exception_info = None
-    result_type = "success"
-
-    def run_test():
-        nonlocal result_type, exception_info
-        try:
-            test(result)
-            if result.failures:
-                result_type = "failure"
-                exception_info = result.failures[0][1]
-            elif result.errors:
-                result_type = "error"
-                exception_info = result.errors[0][1]
-            elif result.skipped:
-                result_type = "skipped"
-                exception_info = result.skipped[0][1]
-        except Exception as e:
-            result_type = "error"
-            exception_info = traceback.format_exc()
-
-    thread = threading.Thread(target=run_test, daemon=True)
-    thread.start()
-    thread.join(timeout=timeout)
-
-    elapsed = time.monotonic() - start_time
-
-    if thread.is_alive():
-        # Test timed out
-        result_type = "timeout"
-        exception_info = f"Test timed out after {timeout} seconds"
-        if log_stream is not None:
-            log_stream.write(f"[PARALLEL] TIMEOUT: {test_id} (after {elapsed:.1f}s)\n")
-            log_stream.flush()
-        # Note: We cannot forcibly kill the thread, but we can continue
-        # The thread will be left as a daemon and cleaned up on exit
-    else:
-        status = result_type.upper()
-        if log_stream is not None:
-            log_stream.write(f"[PARALLEL] Finished: {test_id} [{status}] ({elapsed:.1f}s)\n")
-            log_stream.flush()
-
-    return test, result_type, exception_info, elapsed
-
-
-class ParallelTestResult(unittest.TestResult):
-    """Thread-safe test result that aggregates results from parallel execution."""
-
-    def __init__(self, stream=None):
-        super().__init__()
-        self.stream = stream or sys.stderr
-        self.lock = threading.Lock()
-        self.test_timings = []
-
-    def add_result(self, test, result_type, err_info, elapsed):
-        with self.lock:
-            self.test_timings.append((test.id(), elapsed, result_type.upper()))
-            if result_type == "success":
-                self.successes = getattr(self, "successes", 0) + 1
-            elif result_type == "failure":
-                self.failures.append((test, err_info))
-            elif result_type == "error":
-                self.errors.append((test, err_info))
-            elif result_type == "skipped":
-                self.skipped.append((test, err_info))
-            elif result_type == "timeout":
-                self.errors.append((test, err_info))
-
-
-class ParallelTestRunner:
-    """Run tests in parallel with per-test timeouts."""
-
-    def __init__(self, workers=4, timeout=300, stream=None):
-        self.workers = workers
-        self.timeout = timeout
-        self.stream = stream or sys.stderr
-
-    def run(self, test):
-        tests = list(_flatten_tests(test))
-        result = ParallelTestResult(self.stream)
-        result.testsRun = 0
-
-        self.stream.write(f"[PARALLEL] Running {len(tests)} tests with {self.workers} workers, timeout={self.timeout}s\n")
-        self.stream.flush()
-
-        result.startTestRun()
-        setup_ok_classes = []
-        try:
-            class_errors, setup_ok_classes = _prepare_parallel_class_fixtures(tests, self.stream)
-            runnable_tests = [t for t in tests if t.__class__ not in class_errors]
-            if FPC_RTL_MODE and self.workers > 1:
-                assert runnable_tests, (
-                    "FPC RTL mode enabled but no runnable tests found. "
-                    "Check that KGPC_FPC_RTL=1, FPCSource is present, and test discovery found .expected files."
-                )
-                # Use longest-job-first scheduling: move pp_pas_bootstrap to
-                # the front so it starts early and overlaps with other tests.
-                # pp_pas_bootstrap compiles the entire FPC compiler (~150s) and
-                # is structurally the longest test — update this if renamed.
-                #
-                # No warm-up phase needed: the AST cache uses atomic
-                # write-to-temp-then-rename (PID-unique temp files), so
-                # parallel compiler processes can safely race to populate
-                # the same cache entries without corruption.  The first
-                # writer wins and subsequent processes read the cached
-                # result on the next unit load.  This is faster than
-                # blocking all workers on a sequential warm-up test.
-                reordered = list(runnable_tests)
-                pp_idx = next(
-                    (i for i, t in enumerate(reordered)
-                     if 'pp_pas_bootstrap' in t._testMethodName),
-                    None,
-                )
-                if pp_idx is not None:
-                    reordered.insert(0, reordered.pop(pp_idx))
-                self.stream.write(
-                    f"[PARALLEL] FPC RTL full parallel: "
-                    f"{len(reordered)} tests with {self.workers} workers\n"
-                )
-                self.stream.flush()
-                with concurrent.futures.ThreadPoolExecutor(max_workers=self.workers) as executor:
-                    futures = {
-                        executor.submit(
-                            _run_single_test_with_timeout, t,
-                            getattr(getattr(t, t._testMethodName, None), '_timeout', self.timeout),
-                            self.stream,
-                        ): t
-                        for t in reordered
-                    }
-                    for future in concurrent.futures.as_completed(futures):
-                        try:
-                            test_case, result_type, err_info, elapsed = future.result()
-                            result.add_result(test_case, result_type, err_info, elapsed)
-                            result.testsRun += 1
-                        except Exception:
-                            test_case = futures[future]
-                            result.errors.append((test_case, traceback.format_exc()))
-                            result.testsRun += 1
-            else:
-                with concurrent.futures.ThreadPoolExecutor(max_workers=self.workers) as executor:
-                    futures = {
-                        executor.submit(
-                            _run_single_test_with_timeout, t,
-                            getattr(getattr(t, t._testMethodName, None), '_timeout', self.timeout),
-                            self.stream,
-                        ): t
-                        for t in runnable_tests
-                    }
-                    for future in concurrent.futures.as_completed(futures):
-                        try:
-                            test_case, result_type, err_info, elapsed = future.result()
-                            result.add_result(test_case, result_type, err_info, elapsed)
-                            result.testsRun += 1
-                        except Exception:
-                            test_case = futures[future]
-                            result.errors.append((test_case, traceback.format_exc()))
-                            result.testsRun += 1
-            for t in tests:
-                if t.__class__ in class_errors:
-                    result.add_result(t, "error", class_errors[t.__class__], 0.0)
-        finally:
-            _cleanup_parallel_class_fixtures(setup_ok_classes, self.stream)
-            result.stopTestRun()
-
-        # Print summary
-        self.stream.write(f"\n[PARALLEL] Completed: {result.testsRun} tests\n")
-        self.stream.write(f"[PARALLEL] Failures: {len(result.failures)}, Errors: {len(result.errors)}, Skipped: {len(result.skipped)}\n")
-        self.stream.flush()
-
-        return result
-
-
-class TAPParallelTestResult(unittest.TestResult):
-    """TAP-compatible aggregate result for parallel execution."""
-
-    def __init__(self):
-        super().__init__()
-        self.test_timings = []
-
-    def add_result(self, test, result_type, err_info, elapsed):
-        self.testsRun += 1
-        self.test_timings.append((test.id(), elapsed, result_type.upper()))
-        if result_type == "success":
-            return
-        if result_type == "failure":
-            self.failures.append((test, err_info))
-            return
-        if result_type == "error" or result_type == "timeout":
-            self.errors.append((test, err_info))
-            return
-        if result_type == "skipped":
-            self.skipped.append((test, err_info))
-
-
-class TAPParallelTestRunner:
-    """Run tests in parallel and emit TAP output deterministically."""
-
-    def __init__(self, workers=4, timeout=300, stream=None, failfast=False):
-        self.workers = workers
-        self.timeout = timeout
-        self.stream = stream or sys.stdout
-        self.failfast = failfast
-
-    def _emit(self, line):
-        self.stream.write(f"{line}\n")
-        self.stream.flush()
-
-    def _emit_diagnostic(self, text):
-        for raw_line in str(text).rstrip().splitlines():
-            self._emit(f"# {raw_line}")
-
-    def _emit_tap_result(self, tap_index, test_obj, result_type, err_info, result):
-        result.add_result(test_obj, result_type, err_info, 0.0)
-        test_name = test_obj.id()
-        if result_type == "success":
-            self._emit(f"ok {tap_index} - {test_name}")
-        elif result_type == "skipped":
-            self._emit(f"ok {tap_index} - {test_name} # SKIP {err_info}")
-        else:
-            self._emit(f"not ok {tap_index} - {test_name}")
-            label = "Failure" if result_type == "failure" else "Error"
-            self._emit_diagnostic(f"{label}:")
-            if err_info is not None:
-                self._emit_diagnostic(err_info)
-
-    def run(self, test):
-        tests = list(_flatten_tests(test))
-        result = TAPParallelTestResult()
-        result.startTestRun()
-        self._emit(f"1..{len(tests)}")
-        effective_workers = max(1, min(self.workers, TAP_MAX_WORKERS))
-
-        completed = {}
-        next_to_emit = 0  # next index to emit in order
-        setup_ok_classes = []
-        try:
-            class_errors, setup_ok_classes = _prepare_parallel_class_fixtures(tests)
-
-            # Pre-populate class errors
-            for idx, t in enumerate(tests):
-                if t.__class__ in class_errors:
-                    completed[idx] = (t, "error", class_errors[t.__class__], 0.0)
-
-            runnable = [(idx, t) for idx, t in enumerate(tests) if t.__class__ not in class_errors]
-            if FPC_RTL_MODE and effective_workers > 1:
-                assert runnable, (
-                    "FPC RTL mode enabled but no runnable tests found. "
-                    "Check that KGPC_FPC_RTL=1, FPCSource is present, and test discovery found .expected files."
-                )
-                # Use longest-job-first scheduling: move pp_pas_bootstrap to
-                # the front so it starts early and overlaps with other tests.
-                # pp_pas_bootstrap compiles the entire FPC compiler (~150s) and
-                # is structurally the longest test — update this if renamed.
-                #
-                # No warm-up phase needed: the AST cache uses atomic
-                # write-to-temp-then-rename (PID-unique temp files), so
-                # parallel compiler processes can safely race to populate
-                # the same cache entries without corruption.  The first
-                # writer wins and subsequent processes read the cached
-                # result on the next unit load.  This is faster than
-                # blocking all workers on a sequential warm-up test.
-                reordered = list(runnable)
-                pp_idx = next(
-                    (i for i, (_, t) in enumerate(reordered)
-                     if 'pp_pas_bootstrap' in t._testMethodName),
-                    None,
-                )
-                if pp_idx is not None:
-                    reordered.insert(0, reordered.pop(pp_idx))
-                self._emit(
-                    f"# FPC RTL full parallel: "
-                    f"{len(reordered)} tests with {effective_workers} workers"
-                )
-                with concurrent.futures.ThreadPoolExecutor(max_workers=effective_workers) as executor:
-                    future_to_index = {
-                        executor.submit(
-                            _run_single_test_with_timeout, t,
-                            getattr(getattr(t, t._testMethodName, None), '_timeout', self.timeout),
-                        ): idx
-                        for idx, t in reordered
-                    }
-                    for future in concurrent.futures.as_completed(future_to_index):
-                        idx = future_to_index[future]
-                        test_case = tests[idx]
-                        try:
-                            completed[idx] = future.result()
-                        except Exception:
-                            completed[idx] = (
-                                test_case,
-                                "error",
-                                traceback.format_exc(),
-                                0.0,
-                            )
-                        stop_now = False
-                        while next_to_emit in completed:
-                            test_obj, result_type, err_info, elapsed = completed[next_to_emit]
-                            self._emit_tap_result(next_to_emit + 1, test_obj, result_type, err_info, result)
-                            del completed[next_to_emit]
-                            next_to_emit += 1
-                            if self.failfast and result_type not in ("success", "skipped"):
-                                stop_now = True
-                                break
-                        if stop_now:
-                            for f in future_to_index:
-                                f.cancel()
-                            break
-            else:
-                with concurrent.futures.ThreadPoolExecutor(max_workers=effective_workers) as executor:
-                    future_to_index = {
-                        executor.submit(
-                            _run_single_test_with_timeout, t,
-                            getattr(getattr(t, t._testMethodName, None), '_timeout', self.timeout),
-                        ): idx
-                        for idx, t in runnable
-                    }
-                    for future in concurrent.futures.as_completed(future_to_index):
-                        idx = future_to_index[future]
-                        test_case = tests[idx]
-                        try:
-                            completed[idx] = future.result()
-                        except Exception:
-                            completed[idx] = (
-                                test_case,
-                                "error",
-                                traceback.format_exc(),
-                                0.0,
-                            )
-                        # Stream TAP lines for all consecutive completed tests
-                        while next_to_emit in completed:
-                            test_obj, result_type, err_info, elapsed = completed[next_to_emit]
-                            self._emit_tap_result(next_to_emit + 1, test_obj, result_type, err_info, result)
-                            del completed[next_to_emit]
-                            next_to_emit += 1
-
-            # Emit any remaining results (class errors that were never submitted)
-            while next_to_emit < len(tests):
-                if next_to_emit in completed:
-                    test_obj, result_type, err_info, elapsed = completed[next_to_emit]
-                else:
-                    test_obj = tests[next_to_emit]
-                    result_type = "error"
-                    err_info = "Internal error: missing parallel test result"
-                self._emit_tap_result(next_to_emit + 1, test_obj, result_type, err_info, result)
-                next_to_emit += 1
-        finally:
-            _cleanup_parallel_class_fixtures(setup_ok_classes)
-            result.stopTestRun()
-        return result
-
-
-def read_file_content(filepath):
-    """Reads and returns the content of a file."""
-    with open(filepath, "r") as f:
-        return f.read()
-
 
 class TestCompiler(unittest.TestCase):
